data ahora funciona internamente con doubles, no con floats se mantiene la compatibilidad con el cliente con polimorfismo para las funciones con parametros float*, (setnewdata, setmoredata) para la que devolvia float*, se ha creado una version nueva double* getlastposd(); pasada toda la parte del server de float a doubles

+++ trunk/server/TPFC_device.C
@@ -35,7 +35,7 @@ void TPFC_device::report(){
 
     TrackingPFC_data::datachunk* d = data->getlastdata();
     vrpn_gettimeofday(&current_time, NULL);
-    const float* aux;
+    const double* aux;
     for (int i =0; i<d->size() && i<sensors;i++){
       aux= d->getdata(i);
       position[0]=aux[0];
